to mark

+++ athena/src/main/java/project/dao/bulletproof/BulletproofStatement.java
@@ -1,13 +1,17 @@
 package project.dao.bulletproof;
 
+import project.elgamal.ElGamalPK;
+
+import java.math.BigInteger;
+
 public class BulletproofStatement {
     public final int n;
-    public final BigInteger V; // commitment 
+    public final BigInteger V; // commitment
     public final ElGamalPK pk;
 
 
 
-    public BulletproofStatement(BigInteger n, BigInteger V, ElGamalPK pk) {
+    public BulletproofStatement(int n, BigInteger V, ElGamalPK pk) {
         this.n = n;
         this.V = V;
         this.pk = pk;
